+++ FOX/ff/lj_calculate.py
@@ -52,7 +52,8 @@ def get_non_bonded(mol: Union[str, MultiMolecule],
                    cp2k_settings: Optional[Mapping] = None,
                    max_array_size: int = 10**8,
                    distance_upper_bound: float = np.inf, k: int = 20,
-                   atom_pairs: Optional[Iterable[Tuple[str, str]]] = None) -> pd.DataFrame:
+                   atom_pairs: Optional[Iterable[Tuple[str, str]]] = None
+                   ) -> Tuple[pd.DataFrame, pd.DataFrame]:
     r"""Collect forcefield parameters and calculate all non-covalent interactions in **mol**.
 
     Forcefield parameters (*i.e.* charges and Lennard-Jones :math:`\sigma` and
@@ -105,11 +106,10 @@ def get_non_bonded(mol: Union[str, MultiMolecule],
 
     Returns
     -------
-    :class:`pandas.DataFrame`
-        A DataFrame with the electrostatic and Lennard-Jones components of the
+    :class:`pandas.DataFrame` & :class:`pandas.DataFrame`
+        Two DataFrames with, respectivelly, the electrostatic and Lennard-Jones components of the
         (inter-ligand) potential energy per atom-pair.
-        The potential energy is summed over atoms with matching atom types and
-        averaged over all molecules within **mol**.
+        The potential energy is summed over atoms with matching atom types.
         Units are in atomic units.
 
     See Also
@@ -156,14 +156,17 @@ def get_non_bonded(mol: Union[str, MultiMolecule],
     # Calculate and return the potential energies
     core_atoms = set(psf.atom_type[psf.residue_id == 1])
     ligand_count = psf.residue_id.max() - 1
-    return get_V(mol, slice_dict, prm_df.loc, ligand_count, core_atoms=core_atoms)
+    return get_V(mol, slice_dict, prm_df.loc, ligand_count,
+                 k=k, core_atoms=core_atoms,
+                 max_array_size=max_array_size,
+                 distance_upper_bound=distance_upper_bound)
 
 
 def get_V(mol: MultiMolecule, slice_mapping: SliceMapping,
           prm_mapping: PrmMapping, ligand_count: int,
           core_atoms: Optional[Iterable[str]] = None,
           max_array_size: int = 10**8,
-          distance_upper_bound: float = np.inf, k: int = 20) -> pd.DataFrame:
+          distance_upper_bound: float = np.inf, k: int = 20) -> Tuple[pd.DataFrame, pd.DataFrame]:
     r"""Calculate all non-covalent interactions averaged over all molecules in **mol**.
 
     Parameters
@@ -199,22 +202,21 @@ def get_V(mol: MultiMolecule, slice_mapping: SliceMapping,
 
     Returns
     -------
-    :class:`pandas.DataFrame`
-        A DataFrame with the electrostatic and Lennard-Jones components of the
+    :class:`pandas.DataFrame` & :class:`pandas.DataFrame`
+        Two DataFrames with, respectivelly, the electrostatic and Lennard-Jones components of the
         (inter-ligand) potential energy per atom-pair.
-        The potential energy is summed over atoms with matching atom types and
-        averaged over all molecules within **mol**.
+        The potential energy is summed over atoms with matching atom types.
         Units are in atomic units.
 
     """
     core_atoms = set(core_atoms) if core_atoms is not None else set()
     mol = mol * Units.conversion_ratio('Angstrom', 'au')
 
-    df = pd.DataFrame(
-        0.0,
-        index=pd.MultiIndex.from_tuples(sorted(slice_mapping.keys())),
-        columns=pd.Index(['elstat', 'lj'], name='au')
-    )
+    index = pd.RangeIndex(0, len(mol), name='MD Iteration')
+    columns = pd.MultiIndex.from_tuples(sorted(slice_mapping.keys()), names=['atom1', 'atom2'])
+
+    elstat_df = pd.DataFrame(0.0, index=index.copy(), columns=columns.copy())
+    lj_df = pd.DataFrame(0.0, index=index, columns=columns)
 
     # Specify the function for calculating the distance matrices
     if np.isinf(distance_upper_bound):
@@ -235,15 +237,15 @@ def get_V(mol: MultiMolecule, slice_mapping: SliceMapping,
         # Construct the distance matrices and calculate the potential energies
         for mol_subset in slice_iterator:
             dist = dist_func(mol, ij, ligand_count, contains_core, mol_subset=mol_subset)
-            df.at[atoms, 'elstat'] += get_V_elstat(charge, dist)
-            df.at[atoms, 'lj'] += get_V_lj(sigma, epsilon, dist)
+            elstat_df.loc[elstat_df.index[mol_subset], atoms] = get_V_elstat(charge, dist)
+            lj_df.loc[lj_df.index[mol_subset], atoms] = get_V_lj(sigma, epsilon, dist)
             del dist
 
         if atoms[0] == atoms[1]:  # Avoid double-counting
-            df.loc[atoms] /= 2
+            elstat_df[atoms] /= 2
+            lj_df[atoms] /= 2
 
-    df /= len(mol)
-    return df
+    return elstat_df, lj_df
 
 
 def _get_dist(mol: MultiMolecule, ij: np.ndarray, ligand_count: int,
@@ -275,7 +277,7 @@ def _get_kd_dist(mol: MultiMolecule, ij: np.ndarray, ligand_count: int,
         tree = cKDTree(xyz2)
         dist[n], idx[n] = tree.query(xyz1, k=k, distance_upper_bound=distance_upper_bound)
 
-    # Set all inter-ligand interaction to np.nan
+    # Set all intra-ligand interaction to np.nan
     if not contains_core:
         lig_len = len(i_ar) // ligand_count
 
@@ -307,7 +309,7 @@ def _get_slice_iterator(stop: int, dmat_size: int,
         start += step
 
 
-def get_V_elstat(q: float, dist: np.ndarray) -> float:
+def get_V_elstat(q: float, dist: np.ndarray) -> np.ndarray:
     r"""Calculate and sum the electrostatic potential energy given a distance matrix **dist**..
 
     .. math::
@@ -329,13 +331,16 @@ def get_V_elstat(q: float, dist: np.ndarray) -> float:
     Returns
     -------
     :class:`float`
-        The elctrostatic potential energy summed over all distance pairs in **dist**.
+        The elctrostatic potential energy summed over all distance pairs in **dist** along
+        axises ``>= 1``.
 
     """
-    return np.nansum(q / dist)
+    ret = q / dist
+    axis = tuple(range(1, ret.ndim))
+    return np.nansum(ret, axis=axis)
 
 
-def get_V_lj(sigma: float, epsilon: float, dist: np.ndarray) -> float:
+def get_V_lj(sigma: float, epsilon: float, dist: np.ndarray) -> np.ndarray:
     r"""Calculate and sum the Lennard-Jones potential given a distance matrix **dist**.
 
     .. math::
@@ -368,11 +373,13 @@ def get_V_lj(sigma: float, epsilon: float, dist: np.ndarray) -> float:
 
     Returns
     -------
-    :class:`float`
-        The Lennard-Jones potential energy summed over all distance pairs in **dist**.
+    :class:`numpy.ndarray`
+        The Lennard-Jones potential energy summed over all distance pairs in **dist** along
+        axises ``>= 1``.
 
     """
     sigma_dist = (sigma / dist)**6
     lj = sigma_dist**2 - sigma_dist
     lj *= epsilon * 4
-    return np.nansum(lj)
+    axis = tuple(range(1, lj.ndim))
+    return np.nansum(lj, axis=axis)